Print the current Node name in anomaly mode
Fixes #33675.

I added a `current_node_name` argument to AnomalyMetadata::print_stack.
This is a mandatory arg because I found only one callsite and making it
a default arg on a virtual function can be confusing.

Test Plan:
- Tested locally:
https://gist.github.com/zou3519/09937387c83efc76e1700374d5c9c9d9
- I don't know how to add a test for this: the message is printed to
stderr but it isn't an exception nor a warning. I considered capturing
the stderr of a subprocess but that seems like asking for flakiness.

ghstack-source-id: 7f5f18c4ddd8bb2f15e3f3eb5d21ab0c7923d074
Pull Request resolved: https://github.com/pytorch/pytorch/pull/33875