AccumulateGrad: ensure sparse tensor indices and values refcount is always 1
We check the use_count for indices and values when we avoid a clone
for sparse tensors. The sparse tensor grad itself might have a higher refcount
due to DDP hooks/dist autograd structures holding refs, but the indices and
values inside the sparse tensor should always have a refcount of 1.

Differential Revision: [D20375239](https://our.internmc.facebook.com/intern/diff/D20375239/)

ghstack-source-id: 99900534
Pull Request resolved: https://github.com/pytorch/pytorch/pull/34559